requirements.txt
Django==4.2
Pillow==9.5.0
psycopg2==2.9.6
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0
djangorestframework==3.14.0
mock==5.1.0
